+++ src/components/menus/LoginMenu.svelte
@@ -2,12 +2,11 @@
     import { t } from 'svelte-i18n';
     import { slide } from 'svelte/transition';
     import { cubicOut } from 'svelte/easing';
-    import AuthService from '../../service/user/auth.service';
+    import { authService } from '../../service/user/auth.service';
     import { showError } from '../../store/notification.store';
     import { currentMenu } from '../../store/app.store';
     
     const slideParams = { duration: 200, easing: cubicOut };
-    const authService = new AuthService();
     let email = $state(import.meta.env.DEV ? 'john.doe@example.com' : '');
     let password = $state(import.meta.env.DEV ? 'john.doe@example.com' : '');
     let loginMenu = $state(false);